Prevent enqueueing job if the job has been archived
Previously we'd enqueue a duplicate record in the following sequence of events:

1) Client sends enqueue request. The request is received by the server, but the
   client times out before it receives a reply.
2) Server processes the event, places the job in the archived_jobs table, and
   deletes from the queued_jobs table.
3) Client sleeps and retries. Since the job is not present in the queued_jobs
   table, the server enqueues the job.

This roughly doubles the cost of enqueue queries, from 4 to 8 in an EXPLAIN,
since we have to do an additional SELECT against the archived_jobs table to
avoid the enqueue in that case. We pretty much have to do this though to avoid
inconsistent results in the database.

The good news is, since we insert into archived_jobs before deleting from
queued_jobs, it'll be difficult to lose a race where we read an empty record in
archived_jobs, then successfully try the insert. The following sequence would
have to happen

1. WHERE NOT EXISTS... in Enqueue returns that there are no records in archived_jobs
2. Status callback inserts archived job record
3. Status callback deletes record from queued_jobs
4. INSERT INTO .. completes.

(1) and (4) are completed in the same query to the database, (2) and (3) are
two separate roundtrips to the database, so this seems unlikely.