Support marking jobs as failed & unretryable
Adds a new "retryable" argument to HandleStatusCallback to indicate that the
failure is not retryable, and the job shouldn't be re-enqueued. Adds a test at
the services and at the controller level to test this. Moves the record status
endpoint to its own file, and adds a basic test for it.

Fixes #1.