Stabilize tests part 1
- Restrict tests to Go 1.9 so we can use t.Helper.

- Avoid using DefaultServeMux, which panics when you register two
handlers for the same route (in the case for example that we run the
same test twice in a row in the same process).

- Use httptest.NewRequest, which panics if you get an error, and elide
the _ error handling.

- Fix data race in the SharedSecretAuthorizer (we were not guarding
a map access with a lock) by using a read/write lock.

- Bump test timeouts to account for test parallelization and
a potentially slow database.

- Modify tests that hit the database so they can run concurrently
without interfering with each other. Ensure that we are properly
cleaning up after tests that run in parallel.